pyproject.toml
[tool.poetry]
name = "platform-sdet"
version = "0.1.0"
description = ""
authors = ["Matthew"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.3"
websockets = "^10.0"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
httpx = "^0.27.2"
docker = "^7.1.0"
python-dotenv = "^1.0.1"
allure-pytest = "^2.13.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
